refactor(deps): move RelativeTimeFormat abstract operations to ecma402-abstract
Move 6 ECMA-402 Section 17 abstract operations from
intl-relativetimeformat/abstract/ to ecma402-abstract/RelativeTimeFormat/:
- FormatRelativeTime, FormatRelativeTimeToParts
- InitializeRelativeTimeFormat, MakePartsList
- PartitionRelativeTimePattern, SingularRelativeTimeUnit
