+++ packages/ecma402-abstract/RelativeTimeFormat/BUILD.bazel
@@ -1,0 +1,26 @@
+load("//tools:index.bzl", "generate_ide_tsconfig_json", "ts_compile")
+
+PACKAGE_NAME = "RelativeTimeFormat"
+
+SRCS = glob(["*.ts"])
+
+SRC_DEPS = [
+    "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
+]
+
+ts_compile(
+    name = PACKAGE_NAME,
+    srcs = SRCS,
+    package_json = False,
+    visibility = ["//visibility:public"],
+    deps = SRC_DEPS,
+)
+
+generate_ide_tsconfig_json(
+    composite = True,
+    project_references = [
+        "//packages/ecma402-abstract",
+        "//packages/ecma402-abstract/types",
+    ],
+)
