+++ packages/ecma402-abstract/RelativeTimeFormat/SingularRelativeTimeUnit.ts
