+++ packages/ecma402-abstract/tsconfig.json
@@ -8,6 +8,7 @@
   "exclude": [
     "./DisplayNames/**/*",
     "./NumberFormat/**/*",
+    "./RelativeTimeFormat/**/*",
     "./scripts/**/*",
     "./types/**/*"
   ],
